Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/sandboxed-iframe-expected.txt

CONSOLE MESSAGE: Origin null is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8800/fetch/api/resources/top.txt due to access control checks.


PASS CORS with sandboxed iframe 
